# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UMKM Website API",
        "description": "API Documentation for UMKM Website",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/businesses/{business}/daily-sales": {
            "get": {
                "tags": [
                    "Daily Sale"
                ],
                "summary": "Get daily sales history for a specific business",
                "operationId": "07d6158dec3c1c920a13bd4af8bce172",
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "description": "Business ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start Date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End Date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of daily sales",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Business not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Daily Sale"
                ],
                "summary": "Input daily sales data",
                "operationId": "a978a80b350e81455e95fd832e1726a4",
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "description": "Business ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sale_date",
                                    "items"
                                ],
                                "properties": {
                                    "sale_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-09"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Hujan deras, toko sepi"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "menu_item_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "qty_sold": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 50
                                                },
                                                "qty_produced": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 60
                                                },
                                                "qty_wasted": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Daily sales recorded successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Business not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/businesses/{business}/daily-sales/{sale}": {
            "get": {
                "tags": [
                    "Daily Sale"
                ],
                "summary": "Get specific daily sales detail",
                "operationId": "a94f21af7e6a8d9fedd187fd5b1bc1a1",
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sale",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daily sales detail"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Daily Sale"
                ],
                "summary": "Update specific daily sales data",
                "operationId": "09e51b97469f3d786d7c9bfb8dc7d5f2",
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sale",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "notes": {
                                        "type": "string",
                                        "example": "Updated notes"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "menu_item_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "qty_sold": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 45
                                                },
                                                "qty_produced": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 50
                                                },
                                                "qty_wasted": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Daily sales updated successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/businesses/{business}/daily-sales/import": {
            "post": {
                "tags": [
                    "Daily Sale"
                ],
                "summary": "Import historical daily sales from Excel or CSV file",
                "operationId": "83febbf7027b6a83092220583aa9200e",
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "description": "Business ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "Excel or CSV file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sales data imported successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Business not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}